Fix inefficiency in databus poll due to unclaiming events (#98)
* Added efficiency by uncaching empty markers and rewinding cursors when databus events are renewed with TTL zero

* Changed databus poll to stream results and unclaim events less frequently

* Fixes API key delete issue causing gatekeeper failure

+++ web/src/main/java/com/bazaarvoice/emodb/web/auth/ApiKeyAdminTask.java
@@ -358,14 +358,16 @@ public class ApiKeyAdminTask extends Task {
 
         String key = getValueFromParams("key", parameters);
         ApiKey apiKey = _authIdentityManager.getIdentity(key);
-        checkArgument(apiKey != null, "Unknown API key");
 
-        // Does the caller have permission to revoke every role from the API key?
-        for (String role : apiKey.getRoles()) {
-            subject.checkPermission(Permissions.grantRole(RoleIdentifier.fromString(role)));
+        if (apiKey != null) {
+            // Does the caller have permission to revoke every role from the API key?
+            for (String role : apiKey.getRoles()) {
+                subject.checkPermission(Permissions.grantRole(RoleIdentifier.fromString(role)));
+            }
+
+            _authIdentityManager.deleteIdentity(key);
         }
-        
-        _authIdentityManager.deleteIdentity(key);
+
         output.println("API key deleted");
     }
